+++ CHANGELOG.md
@@ -6,7 +6,7 @@ title: Changelog
 
 ## Unreleased (????-??-??)
 
-...
+* `<cc-pricing-table>`: add new `temporality` type `1000-minutes`
 
 ## 7.3.0 (2021-10-07)